Splitting aux.py into aux_dst.py and aux_kp.py
